% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_hypsometric_model.R
\name{best_hypsometric_model}
\alias{best_hypsometric_model}
\title{Get the best hypsometric model.}
\usage{
best_hypsometric_model(fit_hypsometric_list, criteria, model_name)
}
\arguments{
\item{fit_hypsometric_list}{list resulting from the \code{fit_hypsometric} function.}

\item{criteria}{criteria for selecting the best model. Valid options are \code{syx}, \code{adj_r_squared} or \code{both}. Optional. Default is \code{syx}. Check the details for more information.}

\item{model_name}{force the selection of a specific model. Character. Optional. Check the details for more information.}
}
\value{
a list containing three elements. The element \code{name} contains the name of the selected model. The second element, named \code{model} is the \code{lm} object. The third, named \code{transform}, is a string to be used in the \code{predict_h} function.
}
\description{
Get the best model from a list generated from the \code{fit_hypsometric} function.
}
\details{
When the selection criterion is \code{syx}, the model that resulted in the smallest standard error of the estimates will be selected. If the criterion selected by \code{adj_r_squared}, the selected model will be the one that resulted in the highest value for the adjusted coefficient of determination. If the chosen criterion is \code{both}, the best ranked model considering both criteria will be selected.

To force the use of a specific model, declare one of the twelve models in the \code{model_name} argument. Valid options are: Linear', 'Trorey', 'Cubic', 'Prodan', 'Prodan2', 'Assman', 'Henricksen', 'Stoffel', 'Curtis', 'Petterson', 'Naslund' and 'Mean'.
}
