% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_di.R
\name{bi_di}
\alias{bi_di}
\title{Estimate the diameter at a given height based on a fitted Bi taper equation.}
\usage{
bi_di(dbh, h, hi, coef)
}
\arguments{
\item{dbh}{tree diameter at breast height, in centimeters.}

\item{h}{total tree height, in meters.}

\item{hi}{height at which the diameter will be calculated, in meters.}

\item{coef}{numerical vector containing seven coefficients of the Bi taper function.}
}
\value{
a numeric value indicating the diameter at the given height.
}
\description{
Estimates the diameter at a given height of a tree from the diameter at breast height, total height and the coefficients of the Bi taper function.
}
\details{
the Bi variable-form taper function is represented mathematically by the following expression

di ~ dbh * (log(sin((pi/2) * (hi/h)))/(log(sin((pi/2) * (1.3/h)))))^(b0 + b1 * sin((pi/2) * (hi/h)) + b2 * cos((3 * pi/2) * (hi/h)) + b3 * (sin((pi/2) * (hi/h))/(hi/h)) + b4 * dbh + b5 * (hi/h) * dbh^0.5 + b6 * (hi/h) * h^0.5)
}
