% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hypso.R
\name{fit_hypso}
\alias{fit_hypso}
\title{Fit of hypsometric models.}
\usage{
fit_hypso(data, h_col, dbh_col)
}
\arguments{
\item{data}{tree measurement data.}

\item{h_col}{the name of the column containing the height values. Character.}

\item{dbh_col}{the name of the column containing the diameters at breast height. Character.}
}
\value{
a list containing four elements. The first one, named \code{model_evaluation} is a data.frame containing twelve rows with the names, formulas and evaluation parameters of the fitted models. The second, named \code{models}, is a list containing the twelve \code{lm} models. The third, named \code{curves}, is a data.frame containing the heights predicted by each of the twelve fitted models, as well as the observed dbh values and the name of the models. The last, named \code{obs_data} is a data.frame with two columns referring to the observed values of dbh (\code{dbh_col}) and height (\code{h_col}).
}
\description{
Fit and evaluation of twelve hypsometric models.
}
\details{
Twelve models are fitted and evaluated using this function. The models are 'Linear', 'Trorey', 'Cubic', 'Prodan', 'Prodan2', 'Assman', 'Henricksen', 'Stoffel', 'Curtis', 'Petterson', 'Naslund' and 'Mean'.

The fit statistics calculated to evaluate the models are the adjusted R² (\code{adj_r_squared}) and the standard error of the estimates in percentage (\code{syx_perc}).
}
