% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kozak_di.R
\name{kozak_di}
\alias{kozak_di}
\title{Estimate the diameter at a given height based on a fitted Kozak (2004) taper equation.}
\usage{
kozak_di(dbh, h, hi, coef, p)
}
\arguments{
\item{dbh}{tree diameter at breast height, in centimeters.}

\item{h}{total tree height, in meters.}

\item{hi}{height at which the diameter will be calculated, in meters.}

\item{coef}{numerical vector containing nine coefficients of the Kozak taper function.}

\item{p}{numerical value representing the first inflection point calculated in the segmented model of Max and Burkhart (1976).}
}
\value{
a numeric value indicating the diameter at the given height.
}
\description{
Estimates the diameter at a given height of a tree from the diameter at breast height, total height and the coefficients of the Kozak (2004) taper function.
}
\details{
the Kozak (2004) variable-form taper function is represented mathematically by the following expression

di ~b0*(dbh^b1)\emph{(h^b2)}((1-hih^(1/4))/(1-(p^(1/3))))^(b3\emph{hih^4+b4}(1/exp(dbh/h))+b5*((1-hih^(1/4))/(1-(p^(1/3))))^0.1+b6*(1/dbh)+b7*(h^(1-hih^(1/3)))+b8*((1-hih^(1/4))/(1-(p^(1/3)))))
}
