% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hypsometric.R
\name{plot_hypsometric}
\alias{plot_hypsometric}
\title{Plot the curves generated by the hypsometric models}
\usage{
plot_hypsometric(fit_hypsometric_list)
}
\arguments{
\item{fit_hypsometric_list}{list resulting from the \code{fit_hypsometric} function.}
}
\value{
a list containing two elements. The first one, named \code{model} is the selected model - a \code{lm} object. The second, named \code{transform}, is a string to be used in the \code{predict_h} function.
}
\description{
Plot the curves generated by the fit of the twelve hypsometric models of the \code{fit_hypsometric} function, with the observed data and goodness-of-fit statistics.
}
