% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poli5_hi.R
\name{poli5_hi}
\alias{poli5_hi}
\title{Estimate the height at which a given diameter occurs in a tree, based on a 5th degree polynomial function.}
\usage{
poli5_hi(dbh, h, di, coef)
}
\arguments{
\item{dbh}{tree diameter at breast height, in centimeters.}

\item{h}{total tree height, in meters.}

\item{di}{diameter whose height of occurrence will be estimated, in centimeters.}

\item{coef}{numerical vector containing six coefficients of the 5th degree polynomial function that describes the tree's taper.}
}
\description{
Estimates the height at which a given diameter occurs in a tree, from the diameter at breast height, total height and coefficients of the 5th degree polynomial function that describes the tree's taper.
}
