% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poli5_logs.R
\name{poli5_logs}
\alias{poli5_logs}
\title{Simulate log extraction using a 5th degree polynomial that describes the taper of the tree.}
\usage{
poli5_logs(
  dbh,
  h,
  coef,
  stump_height,
  assortments,
  downgrade,
  eliminate,
  broken,
  defect_height
)
}
\arguments{
\item{dbh}{tree diameter at breast height, in centimeters.}

\item{h}{total tree height, in meters.}

\item{coef}{numerical vector containing six coefficients of the 5th degree polynomial function that describes the tree's taper.}

\item{hi}{final height of the tree section whose volume will be calculated, in meters. Default is the total tree height (h).}

\item{h0}{initial height of the tree section whose volume will be calculated, in meters. Default is 0 (ground height).}
}
\description{
Simulate the extraction of logs from a tree from its measurements, taper function (5th degree polynomial), trunk quality characteristics and harvest parameters such as stump height and assortments.
}
