% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poli5_vol.R
\name{poli5_vol}
\alias{poli5_vol}
\title{Estimate the total or partial volume of the tree, based on a 5th degree polynomial function that describes the taper of the tree.}
\usage{
poli5_vol(dbh, h, coef, hi, h0)
}
\arguments{
\item{dbh}{tree diameter at breast height, in centimeters.}

\item{h}{total tree height, in meters.}

\item{coef}{numerical vector containing six coefficients of the 5th degree polynomial function that describes the tree's taper.}

\item{hi}{final height of the tree section whose volume will be calculated, in meters. Default is the total tree height (h).}

\item{h0}{initial height of the tree section whose volume will be calculated, in meters. Default is 0 (ground height).}
}
\description{
Estimates the total or partial volume of the tree from the diameter at breast height, total height, initial section height, final section height and coefficients of the 5th degree polynomial function that describes the tree's taper.
}
