% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_h.R
\name{predict_h}
\alias{predict_h}
\title{Estimate the height of trees.}
\usage{
predict_h(data, model, transform, dbh_col)
}
\arguments{
\item{data}{tree measurement data.}

\item{model}{one of the twelve hypsometric models present in the \code{fit_hypso} function.}

\item{transform}{transformation to be performed on the predicted values. Use the \code{transform} element from the list resulting from the \code{best_hypso_model} function.}

\item{dbh_col}{the name of the column containing the values from diameter to breast height. Character. Required for Naslund and Prodan functions.}
}
\value{
a numerical vector containing the height estimates.
}
\description{
Estimate the height of trees using one of the twelve hypsometric models in the \code{timbeR} package.
}
\details{
check the \code{fit_hypso} function help for more details.
}
