% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_remove.R
\name{sel_remove}
\alias{sel_remove}
\title{Remove unwanted data by selecting it}
\usage{
sel_remove(data, xvar, yvar)
}
\arguments{
\item{data}{a data.frame.}

\item{xvar}{quoted name of the variable to be displayed in the x axis.}

\item{yvar}{quoted name of the variable to be displayed in the y axis.}
}
\value{
the data.frame given to the \code{data} argument, without the selected points.
}
\description{
Delete unwanted records from the dataset (e.g. outliers) by selecting them ia a scatter plot.
}
