% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_scaling-data.R
\docType{data}
\name{tree_scaling}
\alias{tree_scaling}
\title{Tree scaling example data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 136 rows and 5 columns.
}
\usage{
data(tree_scaling)
}
\description{
Diameter (cm) and height (m) measurements along the bole of 8 Pinus taeda trees.
}
\keyword{datasets}
