% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_hi.R
\name{bi_hi}
\alias{bi_hi}
\title{Estimate the height at which a given diameter occurs in a tree, based on a fitted Bi taper equation.}
\usage{
bi_hi(dbh, h, di, coef)
}
\arguments{
\item{dbh}{tree diameter at breast height, in centimeters.}

\item{h}{total tree height, in meters.}

\item{di}{diameter whose height of occurrence will be estimated, in centimeters.}

\item{coef}{numerical vector containing seven coefficients of the Bi taper equation}
}
\value{
as numeric value indicating the height at which the given diameter occurs.
}
\description{
Estimates the height at which a given diameter occurs in a tree, from the diameter at breast height, total height and coefficients of the Bi taper function.
}
\details{
the Bi variable-form taper function is represented mathematically by the following expression

di ~ dbh * (log(sin((pi/2) * (hi/h)))/(log(sin((pi/2) * (1.3/h)))))^(b0 + b1 * sin((pi/2) * (hi/h)) + b2 * cos((3 * pi/2) * (hi/h)) + b3 * (sin((pi/2) * (hi/h))/(hi/h)) + b4 * dbh + b5 * (hi/h) * dbh^0.5 + b6 * (hi/h) * h^0.5)
}
