% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_scaling-data.R
\docType{data}
\name{tree_scaling}
\alias{tree_scaling}
\title{Tree scaling example data}
\format{
A data frame with 136 rows and 5 variables:
\describe{
\item{tree_id}{tree unique id}
\item{dbh}{diameter at breast height, in centimeters}
\item{h}{total tree height, in meters}
\item{hi}{tree section height, in meters}
\item{di}{diameter at the tree section, in centimeters}
}
}
\usage{
data(tree_scaling)
}
\description{
Diameter (cm) and height (m) measurements along the bole of 8 Pinus taeda trees.
}
\keyword{datasets}
